requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
datasets==2.20.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-huggingface==0.0.3
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.83
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.35.10
orjson==3.10.6
packaging==24.1
pandas==2.2.2
peft==0.11.1
Pillow==10.1.0
priority==2.0.0
protobuf==3.20.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.1.99
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sympy==1.12.1
taskgroup==0.0.0a4
tbb==2021.13.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.4
transformers==4.40.0
trl==0.8.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
xformers==0.0.26.post1
xxhash==3.4.1
yarl==1.9.4
